backend/requirements.txt
oracledb>=2.0.0
python-dotenv
fastapi
uvicorn
